RES-2478: Move nupic.research.frameworks.ax to packages
