+++ packages/ax/nupic/research/frameworks/ax/ray_ax/optimize_accuracy.py
@@ -30,24 +30,21 @@ import ray
 import torch
 from ray import tune
 
-from nupic.research.frameworks.backprop_structure.ray_ax.ray_ax_utils import (
+from nupic.research.frameworks.vernon.ray_custom_logger import DEFAULT_LOGGERS
+
+from .ray_ax_utils import (
     AxSearch,
     ax_client_with_explicit_strategy,
-    filter_to_pareto,
     get_ray_trials,
 )
-from nupic.research.vernon.ray_custom_loggers import DEFAULT_LOGGERS
 
 
-def hyperparameter_loss(config, result, alpha):
+def hyperparameter_loss(config, result):
     log_error = np.log(np.maximum(1 - result["mean_accuracy"], 1e-7))
-    log_nz = np.log(np.maximum(result["inference_nz"], 1))
-    return (alpha * log_error) + ((1 - alpha) * log_nz)
+    return log_error
 
 
-def get_frontier_trials(experiment_dir,
-                        parameters,
-                        num_training_iterations):
+def get_best_config(experiment_dir, parameters, num_training_iterations):
     known_trials = get_ray_trials(experiment_dir, parameters)
     known_trials = [
         (config, result)
@@ -59,35 +56,33 @@ def get_frontier_trials(experiment_dir,
         grouped_trials[config].append(result)
     grouped_trials = list(grouped_trials.items())
 
-    frontier_trials = filter_to_pareto(grouped_trials, [
-        lambda config, results: np.mean([1 - result["mean_accuracy"]
-                                         for result in results]),
-        lambda config, results: np.mean([result["inference_nz"]
-                                         for result in results])
-    ])
+    max_trial = None
+    max_acc = -1
+    for trial in grouped_trials:
+        config, results = trial
+        acc = np.mean([result["mean_accuracy"] for result in results])
+        if acc > max_acc:
+            max_trial = trial
+            max_acc = acc
 
-    # Don't waste time on results that are ~no better than random.
-    frontier_trials = [(config, results)
-                       for config, results in frontier_trials
-                       if np.mean([result["mean_accuracy"]
-                                   for result in results]) > 0.2]
-
-    return frontier_trials
+    return max_trial
 
 
 def insert_hyperparameter_loss(trainable):
+    """
+    Create a new Ray Trainable class that inserts the hyperparameter loss into
+    each result.
+    """
+
     class Cls(trainable):
         def _setup(self, config):
-            self.alpha = config["alpha"]
-            config2 = dict(config)
-            del config2["alpha"]
-            self.config = config2
-            super()._setup(config2)
+            self.config = config
+            super()._setup(config)
 
         def _train(self):
             result = super()._train()
-            result["hyperparameter_loss"] = hyperparameter_loss(
-                self.config, result, self.alpha)
+            result["hyperparameter_loss"] = hyperparameter_loss(self.config,
+                                                                result)
             return result
 
     Cls.__name__ = trainable.__name__
@@ -95,23 +90,15 @@ def insert_hyperparameter_loss(trainable):
     return Cls
 
 
-def ax_optimize_accuracy_weightsparsity(
-        exploring_trainable, followup_trainable, experiment_name, script_dir,
-        alphas, parameters, num_training_iterations,
-        samples_per_frontier_config):
+def ax_optimize_accuracy(exploring_trainable, followup_trainable,
+                         experiment_name, script_dir, parameters,
+                         num_training_iterations, num_best_config_samples):
     """
-    Optimize a Ray Trainable's "mean_accuracy" and "inference_nz", using Ax to
-    select hyperparameters. This procedure will pick up any existing results for
-    this experiment and plug them into the Ax model, then will generate new
-    results. To optimize two objectives, this procedure chooses a set of
-    projections from two scalar objectives to a single scalar objective. These
-    projections are defined by alpha*log(error) + (1 - alpha)*log(inference_nz).
-    The procedure loops over the different alphas and rebuilds the Ax model and
-    generates new results for each. All results go into the same Ray experiment
-    folder, so each search (i.e. each alpha) benefits from all of the other
-    searches. Finally, this procedure finds all configurations at the Pareto
-    frontier of the two objectives and gathers additional results for each until
-    every configuration at the frontier is well-sampled.
+    Optimize a Ray Trainable's "mean_accuracy", using Ax to select
+    hyperparameters. This method will pick up any existing results for this
+    experiment and plug them into the Ax model, then will generate new results.
+    Finally, this procedure finds the best configuration and collects additional
+    results for the configuration to ensure that it is well-sampled.
     """
 
     experiment_dir = os.path.expanduser(f"~/ray_results/{experiment_name}")
@@ -119,8 +106,10 @@ def ax_optimize_accuracy_weightsparsity(
     parser = argparse.ArgumentParser()
     parser.add_argument("--run-initial-configs", action="store_true")
     parser.add_argument("--num-iterations", type=int, default=1)
-    parser.add_argument("--num-random", type=int, default=16)
-    parser.add_argument("--num-configs", type=int, default=16)
+    parser.add_argument("--num-random", type=int, default=16,
+                        help="The number of random configs to test.")
+    parser.add_argument("--num-configs", type=int, default=16,
+                        help="The number of non-random configs to test.")
     args = parser.parse_args()
 
     ray.init()
@@ -131,19 +120,18 @@ def ax_optimize_accuracy_weightsparsity(
         if input_file.exists():
             print("STAGE 0: Test initial suggested hyperparameters")
             with open(input_file, "r") as f:
-                frontier_trials = json.load(f)
+                best_trial = json.load(f)
 
-            configs = [dict(config)
-                       for config, _, _ in frontier_trials]
+            best_config, _ = best_trial
+            best_config = dict(best_config)
             # Only include parameters specified in the parameters list.
-            configs = [{parameter["name"]: config[parameter["name"]]
-                        for parameter in parameters}
-                       for config in configs]
+            best_config = {parameter["name"]: best_config[parameter["name"]]
+                           for parameter in parameters}
 
             tune.run(
                 exploring_trainable,
                 name=experiment_name,
-                config=tune.grid_search(configs),
+                config=dict(best_config),
                 num_samples=1,
                 resources_per_trial={
                     "cpu": 1,
@@ -155,11 +143,8 @@ def ax_optimize_accuracy_weightsparsity(
     num_random_remaining = args.num_random
 
     for _ in range(args.num_iterations):
-        print("STAGE 1: Hyperparameter search")
-        for alpha in alphas:
-            if num_random_remaining + args.num_configs == 0:
-                break
-
+        if num_random_remaining + args.num_configs > 0:
+            print("STAGE 1: Hyperparameter search")
             known_trials = get_ray_trials(experiment_dir, parameters)
             known_trials = [
                 (config, result)
@@ -172,7 +157,7 @@ def ax_optimize_accuracy_weightsparsity(
                                         objective_name="hyperparameter_loss",
                                         minimize=True)
             for config, result in known_trials:
-                loss = hyperparameter_loss(config, result, alpha)
+                loss = hyperparameter_loss(config, result)
                 _, trial_index = ax_client.attach_trial(dict(config))
                 ax_client.complete_trial(
                     trial_index=trial_index,
@@ -185,7 +170,6 @@ def ax_optimize_accuracy_weightsparsity(
                                     max_concurrent=(torch.cuda.device_count()
                                                     if torch.cuda.is_available()
                                                     else 1)),
-                config=dict(alpha=alpha),
                 num_samples=args.num_configs + num_random_remaining,
                 resources_per_trial={
                     "cpu": 1,
@@ -196,24 +180,17 @@ def ax_optimize_accuracy_weightsparsity(
 
             num_random_remaining = 0
 
-        # Verify the frontier
-        print("STAGE 2: Make sure we have enough samples for each point at the "
-              "frontier")
+        print("STAGE 2: Make sure we have enough samples for the best config")
         while True:
-            resample_configs = []
-            for config, results in get_frontier_trials(experiment_dir,
-                                                       parameters,
-                                                       num_training_iterations):
-                if len(results) < samples_per_frontier_config:
-                    resample_configs += (samples_per_frontier_config
-                                         - len(results)) * [dict(config)]
-
-            if len(resample_configs) > 0:
+            best_config, best_results = get_best_config(experiment_dir,
+                                                        parameters,
+                                                        num_training_iterations)
+            if len(best_results) < num_best_config_samples:
                 tune.run(
                     followup_trainable,
                     name=experiment_name,
-                    config=tune.grid_search(resample_configs),
-                    num_samples=1,
+                    config=dict(best_config),
+                    num_samples=num_best_config_samples - len(best_results),
                     resources_per_trial={
                         "cpu": 1,
                         "gpu": (1 if torch.cuda.is_available() else 0)},
@@ -223,15 +200,12 @@ def ax_optimize_accuracy_weightsparsity(
             else:
                 break
 
-    output = []
-    for config, results in get_frontier_trials(experiment_dir,
-                                               parameters,
-                                               num_training_iterations):
-        acc = np.mean([result["mean_accuracy"]
-                       for result in results])
-        nz = np.mean([result["inference_nz"]
-                      for result in results])
-        output.append((config, acc, nz))
+    best_config, best_results = get_best_config(experiment_dir,
+                                                parameters,
+                                                num_training_iterations)
+    acc = np.mean([result["mean_accuracy"]
+                   for result in best_results])
+    output = (best_config, acc)
 
     output_dir = Path(script_dir) / "output"
     os.makedirs(output_dir, exist_ok=True)
