+++ setup.cfg
@@ -72,9 +72,6 @@ install_requires =
 #
 imagenet =
     pretrainedmodels==0.7.4
-    ax-platform
-    botorch
-    gpytorch
     pytorch-lightning==1.2.9
 
 gsc =
